+++ packages/lit-element/src/decorators.ts
@@ -21,7 +21,7 @@
 
 export * from './decorators/customElement.js';
 export * from './decorators/property.js';
-export * from './decorators/internalProperty.js';
+export * from './decorators/state.js';
 export * from './decorators/eventOptions.js';
 export * from './decorators/query.js';
 export * from './decorators/queryAll.js';